Change namespaces and healthController

+++ ProjectBeta/Assets/_ProjectBeta/Scripts/IDamageable.cs
@@ -2,7 +2,7 @@ namespace _ProjectBeta.Scripts
 {
     public interface IDamageable
     {
-        void DoDamage(int damage);
+        void DoDamage(float damage);
         
         //void AddEffect(Effect effectOne, Effect effectTwo. . .)
     }
